(refactor: test) 테스트 분리 및 DisplayName 추가
